==> 2.Nebula/1.NEBULA_data.ipynb <==
import numpy as np
import pandas as pd
import scanpy as sc

sc.settings.verbosity = 3             # verbosity: errors (0), warnings (1), info (2), hints (3)
sc.logging.print_header()
sc.settings.set_figure_params(dpi=100, facecolor='white')

data = sc.read_h5ad('../1.Toomanycells_result/Tcell/data/Tcell_PDL1.h5ad')

np.unique(data.obs['Sub_Cluster'])

subcluster = 't_CD8-CXCL13'

adata = data[data.obs['Sub_Cluster']==subcluster]
adata

adata.to_df()

ann = pd.read_csv('../1.Toomanycells_result/Tcell/result/cluster/'+subcluster+'_cluster.csv',
                  index_col=0,header=0,sep=',',
                 usecols=['cell','cluster'])
ann

adata.obs = adata.obs.join(ann)

adata.obs['cluster'] = adata.obs['cluster'].replace((6),0)

# 设置case组和control组
for i in adata.obs_names:
    if adata.obs.loc[i,'cluster']==16:
        adata.obs.loc[i,'divide']='case'
    else:
        adata.obs.loc[i,'divide']='control'

np.unique(adata.obs['divide'])

mb = pd.read_csv('./data/metabolism_genelist_total.csv',header=None)
mb

em = adata[:,adata.var_names.isin(mb[0])].to_df()
em

em = adata.to_df()
em

em=em.T

em.to_csv('./data/'+subcluster+'_nebula_total.csv',
         header=True,index=True,sep=',')

adata.obs[['divide']].to_csv('./data/'+subcluster+'_nebula_label3.csv',
                           header=False, index=False,sep=',')

adata.obs[['Sample']].to_csv('./data/'+subcluster+'_nebula_batch.csv',
                           header=False, index=False,sep=',')
